requirements.txt
fastapi
uvicorn
scikit-learn
sentence-transformers
numpy
faiss-cpu